HW02-main

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryUserMealRepositoryImpl.java
@@ -1,11 +1,13 @@
 package ru.javawebinar.topjava.repository.mock;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.UserMeal;
 import ru.javawebinar.topjava.repository.UserMealRepository;
 import ru.javawebinar.topjava.util.UserMealsUtil;
 
-import java.util.Collection;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 
@@ -13,7 +15,10 @@ import java.util.concurrent.atomic.AtomicInteger;
  * GKislin
  * 15.09.2015.
  */
+@Repository
 public class InMemoryUserMealRepositoryImpl implements UserMealRepository {
+
+    private static final Logger LOG = LoggerFactory.getLogger(InMemoryUserMealRepositoryImpl.class);
     private Map<Integer, UserMeal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
 
@@ -23,6 +28,7 @@ public class InMemoryUserMealRepositoryImpl implements UserMealRepository {
 
     @Override
     public UserMeal save(UserMeal userMeal) {
+        LOG.info("save " + userMeal);
         if (userMeal.isNew()) {
             userMeal.setId(counter.incrementAndGet());
         }
@@ -31,18 +37,33 @@ public class InMemoryUserMealRepositoryImpl implements UserMealRepository {
     }
 
     @Override
-    public void delete(int id) {
-        repository.remove(id);
+    public boolean delete(int id) {
+        LOG.info("delete " + id);
+        if(repository.containsKey(id))
+        {
+            repository.remove(id);
+            return true;
+        }
+        return false;
     }
 
     @Override
     public UserMeal get(int id) {
+        LOG.info("get " + id);
         return repository.get(id);
     }
 
     @Override
-    public Collection<UserMeal> getAll() {
-        return repository.values();
+    public List<UserMeal> getAll(int userId) {
+        LOG.info("getAll");
+        List<UserMeal> sortedUserMeals = new ArrayList<>();
+        repository.values()
+                .stream()
+                .filter(userMeal -> userMeal.isUserBelong(userId))
+                .forEach(userMeal -> sortedUserMeals.add(userMeal));
+        Collections.sort(sortedUserMeals,
+                (userMeal1, userMeal2) -> userMeal1.getDateTime().isBefore(userMeal2.getDateTime()) ? -1 : 1);
+        return sortedUserMeals.isEmpty() ? null : sortedUserMeals;
     }
 }
